+++ cc-service-api/src/main/java/com/jtmc/apps/reforma/api/v1/annotations/JwtUserClaim.java
@@ -3,6 +3,15 @@ package com.jtmc.apps.reforma.api.v1.annotations;
 public class JwtUserClaim {
 
     private String subject;
+    private int id;
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
 
     public void setSubject(String subject) {
         this.subject = subject;